refactor: 💡 flytte tester til test-mappen
